+++ src/Button.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import classNames from 'classnames';
 import elementType from 'react-prop-types/lib/elementType';
+import SafeAnchor from './SafeAnchor';
 
 const types = ['button', 'reset', 'submit'];
 
@@ -9,37 +10,11 @@ import { Sizes, State, DEFAULT, PRIMARY, LINK } from './styleMaps';
 
 const ButtonStyles = State.values().concat(DEFAULT, PRIMARY, LINK);
 
-let Button = React.createClass({
+class Button extends React.Component {
 
-  propTypes: {
-    active: React.PropTypes.bool,
-    disabled: React.PropTypes.bool,
-    block: React.PropTypes.bool,
-    navItem: React.PropTypes.bool,
-    navDropdown: React.PropTypes.bool,
-    /**
-     * You can use a custom element for this component
-     */
-    componentClass: elementType,
-    href: React.PropTypes.string,
-    target: React.PropTypes.string,
-    /**
-     * Defines HTML button type Attribute
-     * @type {("button"|"reset"|"submit")}
-     * @defaultValue 'button'
-     */
-    type: React.PropTypes.oneOf(types)
-  },
-
-  getDefaultProps() {
-    return {
-      active: false,
-      block: false,
-      disabled: false,
-      navItem: false,
-      navDropdown: false
-    };
-  },
+  constructor(props, context) {
+    super(props, context);
+  }
 
   render() {
     let classes = this.props.navDropdown ? {} : bootstrapUtils.getClassSet(this.props);
@@ -61,23 +36,23 @@ let Button = React.createClass({
       'renderAnchor' : 'renderButton';
 
     return this[renderFuncName](classes);
-  },
+  }
 
   renderAnchor(classes) {
-    let Component = this.props.componentClass || 'a';
-    let href = this.props.href || '#';
-    classes.disabled = this.props.disabled;
+    let { disabled, href } = this.props;
+
+    classes.disabled = disabled;
 
     return (
-      <Component
+      <SafeAnchor
         {...this.props}
-        href={href}
+        href={href || '#'}
         className={classNames(this.props.className, classes)}
-        role="button">
+      >
         {this.props.children}
-      </Component>
+      </SafeAnchor>
     );
-  },
+  }
 
   renderButton(classes) {
     let Component = this.props.componentClass || 'button';
@@ -90,7 +65,7 @@ let Button = React.createClass({
         {this.props.children}
       </Component>
     );
-  },
+  }
 
   renderNavItem(classes) {
     let liClasses = {
@@ -103,7 +78,36 @@ let Button = React.createClass({
       </li>
     );
   }
-});
+}
+
+Button.propTypes = {
+  active: React.PropTypes.bool,
+  disabled: React.PropTypes.bool,
+  block: React.PropTypes.bool,
+  navItem: React.PropTypes.bool,
+  navDropdown: React.PropTypes.bool,
+  onClick: React.PropTypes.func,
+  /**
+   * You can use a custom element for this component
+   */
+  componentClass: elementType,
+  href: React.PropTypes.string,
+  target: React.PropTypes.string,
+  /**
+   * Defines HTML button type Attribute
+   * @type {("button"|"reset"|"submit")}
+   * @defaultValue 'button'
+   */
+  type: React.PropTypes.oneOf(types)
+};
+
+Button.defaultProps = {
+  active: false,
+  block: false,
+  disabled: false,
+  navItem: false,
+  navDropdown: false
+};
 
 Button.types = types;