+++ src/Nav.js
@@ -1,19 +1,39 @@
 import React, { cloneElement } from 'react';
+import ReactDOM from 'react-dom';
 import classNames from 'classnames';
 import all from 'react-prop-types/lib/all';
-import deprecated from 'react-prop-types/lib/deprecated';
+import warning from 'warning';
+import keycode from 'keycode';
 import tbsUtils, { bsStyles, bsClass as _bsClass } from './utils/bootstrapUtils';
+import { nextEnabled, TAB, PANE } from './utils/tabUtils';
 
 import ValidComponentChildren from './utils/ValidComponentChildren';
-import createChainedFunction from './utils/createChainedFunction';
-
-import Collapse from './Collapse';
+import chain from './utils/createChainedFunction';
 
 class Nav extends React.Component {
 
+  componentDidUpdate() {
+    if (this._needsRefocus) {
+      let ul = this.refs.ul && ReactDOM.findDOMNode(this.refs.ul);
+      let tabs = ul ? ul.children || [] : [];
+      let tabIdx = this.eventKeys().indexOf(this.getActiveKey());
+
+      this._needsRefocus = false;
+
+      if (tabIdx !== -1) {
+        let tabNode = tabs[tabIdx];
+
+        if (tabNode && tabNode.firstChild) {
+          tabNode.firstChild.focus();
+        }
+      }
+    }
+  }
+
   render() {
-    const { className, ulClassName, id, ulId } = this.props;
+    const { className } = this.props;
     const isNavbar = this.props.navbar != null ? this.props.navbar : this.context.$bs_navbar;
+
     const classes = tbsUtils.getClassSet(this.props);
 
     classes[tbsUtils.prefix(this.props, 'stacked')] = this.props.stacked;
@@ -21,10 +41,9 @@ class Nav extends React.Component {
 
     if (isNavbar) {
       let bsClass = this.context.$bs_navbar_bsClass || 'navbar';
-      const navbarRight = this.props.right != null ? this.props.right : this.props.pullRight;
 
       classes[tbsUtils.prefix({ bsClass }, 'nav')] = true;
-      classes[tbsUtils.prefix({ bsClass }, 'right')] = navbarRight;
+      classes[tbsUtils.prefix({ bsClass }, 'right')] = this.props.pullRight;
       classes[tbsUtils.prefix({ bsClass }, 'left')] = this.props.pullLeft;
     } else {
       classes['pull-right'] = this.props.pullRight;
@@ -34,32 +53,59 @@ class Nav extends React.Component {
     let list = (
       <ul ref="ul"
         {...this.props}
-        id={ulId || id}
-        role={this.props.bsStyle === 'tabs' ? 'tablist' : null}
-        className={classNames(className, ulClassName, classes)}
+        role={this.getNavRole()}
+        className={classNames(className, classes)}
       >
         {ValidComponentChildren.map(this.props.children, this.renderNavItem, this)}
       </ul>
     );
 
-    // TODO remove in 0.29
-    if (this.context.$bs_deprecated_navbar && this.props.collapsible) {
-      list = (
-        <Collapse
-          in={this.props.expanded}
-          className={isNavbar ? 'navbar-collapse' : void 0}
-        >
-          <div>
-            { list }
-          </div>
-        </Collapse>
-      );
-    }
-
     return list;
   }
 
-  getChildActiveProp(child) {
+  renderNavItem(child, index) {
+    let onSelect = chain(child.props.onSelect, this.props.onSelect);
+    let active = this.isChildActive(child);
+    let tabProps = this.getTabProps(child, active, onSelect);
+
+    return cloneElement(
+      child,
+      {
+        active,
+        activeKey: this.props.activeKey,
+        activeHref: this.props.activeHref,
+        onSelect,
+        key: child.key || index,
+        navItem: true,
+        ...tabProps
+      }
+    );
+  }
+
+  getActiveKey() {
+    const context = this.context.$bs_tabcontainer;
+    if (!context) {
+      return this.props.activeKey;
+    }
+
+    warning(!(this.props.activeKey != null || this.props.activeHref),
+      'Specifing a Nav `activeKey` or `activeHref` prop in the context of a `TabContainer` is not supported. ' +
+      'Instead use `<TabContainer activeKey={' + this.props.activeKey + '} />`');
+
+    return context.activeKey;
+  }
+
+  isChildActive(child) {
+    let activeKey = this.getActiveKey();
+
+    if (this.context.$bs_tabcontainer) {
+      warning(!child.props.active,
+        'Specifying a NavItem `active` prop in the context of a `TabContainer` is not supported. Instead ' +
+        'use `<TabContainer activeKey={' + child.props.eventKey + '} />`');
+
+      return child.props.eventKey === activeKey;
+    }
+
     if (child.props.active) {
       return true;
     }
@@ -77,27 +123,115 @@ class Nav extends React.Component {
     return child.props.active;
   }
 
-  renderNavItem(child, index) {
-    return cloneElement(
-      child,
-      {
-        role: this.props.bsStyle === 'tabs' ? 'tab' : null,
-        active: this.getChildActiveProp(child),
-        activeKey: this.props.activeKey,
-        activeHref: this.props.activeHref,
-        onSelect: createChainedFunction(child.props.onSelect, this.props.onSelect),
-        key: child.key ? child.key : index,
-        navItem: true
+  getTabProps(child, isActive, onSelect) {
+    const navRole = this.getNavRole();
+    const context = this.context.$bs_tabcontainer;
+
+    if (!context && navRole !== 'tablist') {
+      // No tab props here.
+      return null;
+    }
+
+    let {
+        linkId
+      , 'aria-controls': controls
+      , eventKey
+      , role
+      , onKeyDown
+      , tabIndex = 0 } = child.props;
+
+    if (context && context.getId) {
+      warning(!(linkId || controls),
+        'In the context of a TabContainer, NavItems are given generated `linkId` and `aria-controls` ' +
+        'attributes for the sake of proper component accessibility. Any provided ones will be ignored. ' +
+        'To control these attributes directly provide a `generateChildId` prop to the parent TabContainer.'
+      );
+
+      linkId = context.getId(eventKey, TAB) || null;
+      controls = context.getId(eventKey, PANE) || null;
+      onSelect = chain(onSelect, context.onSelect);
+    }
+
+    if (navRole === 'tablist') {
+      role = role || 'tab';
+      onKeyDown = chain(
+        this.handleTabKeyDown.bind(this, onSelect || (()=>{})),
+        onKeyDown
+      );
+      tabIndex = isActive ? tabIndex : -1;
+    }
+
+    return {
+      onSelect,
+      linkId,
+      role,
+      onKeyDown,
+      'aria-controls': controls,
+      tabIndex,
+    };
+  }
+
+  handleTabKeyDown(onSelect, event) {
+    let keys = this.eventKeys();
+    let currentKey = this.getActiveKey() || keys[0];
+    let next;
+
+    switch (event.keyCode) {
+
+    case keycode.codes.left:
+    case keycode.codes.up:
+      next = nextEnabled(this.props.children, currentKey, keys, false);
+
+      if (next && next !== currentKey) {
+        event.preventDefault();
+        onSelect(next);
+        this._needsRefocus = true;
       }
-    );
+      break;
+    case keycode.codes.right:
+    case keycode.codes.down:
+      next = nextEnabled(this.props.children, currentKey, keys, true);
+
+      if (next && next !== currentKey) {
+        event.preventDefault();
+        onSelect(next);
+        this._needsRefocus = true;
+      }
+      break;
+    default:
+    }
+  }
+
+  eventKeys() {
+    let keys = [];
+    ValidComponentChildren.forEach(this.props.children,
+      ({props: { eventKey }}) => keys.push(eventKey));
+    return keys;
+  }
+
+  getNavRole() {
+    return this.props.role || (this.context.$bs_tabcontainer ? 'tablist' : null);
   }
 }
 
 Nav.propTypes = {
+
+  /**
+   * Marks the child NavItem with a matching `href` prop as active.
+   */
   activeHref: React.PropTypes.string,
+
+  /**
+   * Marks the NavItem with a matching `eventKey` as active. Has a
+   * higher precedence over `activeHref`.
+   */
   activeKey: React.PropTypes.any,
 
+  /**
+   * NavItems are be positioned vertically.
+   */
   stacked: React.PropTypes.bool,
+
   justified: all(
     React.PropTypes.bool,
     ({justified, navbar}) => (
@@ -105,6 +239,17 @@ Nav.propTypes = {
         Error('justified navbar `Nav`s are not supported') : null
     )
   ),
+
+  /**
+   * A callback fired when a NavItem is selected.
+   *
+   * ```js
+   * function (
+   * 	Any eventKey,
+   * 	SyntheticEvent event?
+   * )
+   * ```
+   */
   onSelect: React.PropTypes.func,
 
   /**
@@ -118,51 +263,44 @@ Nav.propTypes = {
     React.PropTypes.string,
     React.PropTypes.number
   ]),
-  /**
-   * CSS classes for the inner `ul` element
-   *
-   * @deprecated
-   */
-  ulClassName: deprecated(React.PropTypes.string,
-    'The wrapping `<nav>` has been removed you can use `className` now'),
-  /**
-   * HTML id for the inner `ul` element
-   *
-   * @deprecated
-   */
 
-  ulId: deprecated(React.PropTypes.string,
-    'The wrapping `<nav>` has been removed you can use `id` now'),
+  /**
+   * ARIA role for the Nav, in the context of a TabContainer, the default will be set
+   * to "tablist", but can be overridden by the Nav when set explicitly.
+   *
+   * When the role is set to "tablist" NavItem focus is managed according to the
+   * ARIA authoring practices for tabs: https://www.w3.org/TR/2013/WD-wai-aria-practices-20130307/#tabpanel
+   */
+  role: React.PropTypes.string,
 
   /**
    * Apply styling an alignment for use in a Navbar. This prop will be set
    * automatically when the Nav is used inside a Navbar.
    */
   navbar: React.PropTypes.bool,
-  eventKey: React.PropTypes.any,
+
+  /**
+   * Float the Nav to the right. When `navbar` is `true` the appropriate
+   * contextual classes are added as well.
+   */
   pullRight: React.PropTypes.bool,
+
+  /**
+   * Float the Nav to the left. When `navbar` is `true` the appropriate
+   * contextual classes are added as well.
+   */
   pullLeft: React.PropTypes.bool,
-
-  right: deprecated(React.PropTypes.bool,
-    'Use the `pullRight` prop instead'),
-
-  /**
-   * @private
-   */
-  expanded: React.PropTypes.bool,
-
-  /**
-   * @private
-   */
-  collapsible: deprecated(React.PropTypes.bool,
-    'Use `Navbar.Collapse` instead, to create collapsible Navbars'),
 };
 
 Nav.contextTypes = {
   $bs_navbar: React.PropTypes.bool,
   $bs_navbar_bsClass: React.PropTypes.string,
 
-  $bs_deprecated_navbar: React.PropTypes.bool
+  $bs_tabcontainer: React.PropTypes.shape({
+    activeKey: React.PropTypes.any,
+    onSelect: React.PropTypes.func,
+    getId: React.PropTypes.func
+  })
 };
 
 Nav.defaultProps = {