+++ test/GlyphiconSpec.js
@@ -4,6 +4,8 @@ import ReactDOM from 'react-dom';
 
 import Glyphicon from '../src/Glyphicon';
 
+import { shouldWarn } from './helpers';
+
 describe('Glyphicon', () => {
   it('Should have correct class', () => {
     let instance = ReactTestUtils.renderIntoDocument(
@@ -21,8 +23,10 @@ describe('Glyphicon', () => {
     assert.notOk(ReactDOM.findDOMNode(instance).className.match(/\bform-control-feedback\b/));
   });
 
-  context('when setting the formControlFeedback prop', () => {
+  describe('when setting the formControlFeedback prop', () => {
     it('should have the .form-control-feedback class set', () => {
+      shouldWarn('deprecated');
+
       let instance = ReactTestUtils.renderIntoDocument(
         <Glyphicon formControlFeedback glyph='star' />
       );